koniec zadania 4

+++ zadanie 4/Program.cs
@@ -12,9 +12,9 @@ namespace zadanie_4
         {
             double a, b;
             Console.Write("Podaj a=");
-            a = Convert.ToInt32(Console.ReadLine());
+            a = Convert.ToDouble(Console.ReadLine());
             Console.Write("Podaj b=");
-            b = Convert.ToInt32(Console.ReadLine());
+            b = Convert.ToDouble(Console.ReadLine());
             Console.WriteLine("Pole={0}", a * b);
             Console.WriteLine("Obwód={0}", 2 * a + 2 * b);
             Console.ReadKey(true); //pauza
